feat: Add cancel token for darwin
