+++ lib/src/internal/plugin.dart
@@ -12,6 +12,7 @@ import 'package:photo_manager/platform_utils.dart';
 import '../filter/base_filter.dart';
 import '../filter/classical/filter_option_group.dart';
 import '../filter/path_filter.dart';
+import '../types/cancel_token.dart';
 import '../types/entity.dart';
 import '../types/thumbnail.dart';
 import '../types/types.dart';
@@ -21,31 +22,6 @@ import 'enums.dart';
 import 'progress_handler.dart';
 
 PhotoManagerPlugin plugin = PhotoManagerPlugin();
-
-/// The cancel token is used to cancel the request.
-class PMCancelToken {
-  PMCancelToken({this.debugLabel}) : index = getIndex();
-
-  static int _index = 0;
-
-  static int getIndex() {
-    final res = _index;
-    _index++;
-    return res;
-  }
-
-  final String? debugLabel;
-  final int index;
-
-  /// The key of cancel token, usually to use by [PhotoManagerPlugin.cancelRequest].
-  /// User don't need to use this.
-  String get key => _index.toString();
-
-  /// Cancel the request.
-  Future<void> cancelRequest() async {
-    await plugin.cancelRequest(this);
-  }
-}
 
 mixin BasePlugin {
   MethodChannel _channel = const PMMethodChannel(PMConstants.channelPrefix);
